Adjust apikey.Create to make the metadata functional options
